requirements.txt
openai
langchain
streamlit
elevenlabs
speechrecognition
pyaudio
python-dotenv
whisper